added error handling

+++ cbz.py
@@ -4,11 +4,13 @@ from shutil import make_archive
 
 def main(path):
     # Get all the files in the directory
-    chp_name = os.path.basename(path)
-    name = input("What is the name of the comic? ")
-    make_archive(name + ' ' + chp_name, "zip", path)
-    os.rename(name + ' ' + chp_name + '.zip', name + ' ' + chp_name + '.cbz')
-
+    if os.path.exists(path):
+        chp_name = os.path.basename(path)
+        name = input("What is the name of the comic? ")
+        make_archive(name + ' ' + chp_name, "zip", path)
+        os.rename(name + ' ' + chp_name + '.zip', name + ' ' + chp_name + '.cbz')
+    else:
+        print('That is not a valid directory')
 
 if __name__ == '__main__':
     if len(sys.argv) == 2:
